+++ docs/src/components/mod.rs
@@ -1,6 +1,8 @@
 pub mod animated_flower;
+pub mod animations;
 pub mod cube_animation;
 pub mod extras;
 pub mod navbar;
 pub mod page_not_found;
+pub mod page_transition;
 pub mod transformation_example;
